requirements.txt
# === DEPENDENCIAS PRINCIPALES ===
# PyTorch ecosystem
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Hugging Face ecosystem
transformers>=4.35.0
huggingface_hub>=0.19.0
datasets>=2.14.0

# Fine-tuning eficiente
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# Procesamiento de datos
Pillow>=9.0.0
pycocotools>=2.0.6
numpy>=1.21.0,<2.0
pandas>=1.5.0

# Utilidades
requests>=2.28.0
tqdm>=4.64.0
scipy>=1.9.0

# Opcional para debugging y análisis
matplotlib>=3.5.0
seaborn>=0.11.0
